backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
Pillow==10.0.1
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-filter==23.3
celery==5.3.4
redis==5.0.1